Refactor FilterModal into separate component
